backend/server/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
playwright==1.30.0
reportlab==3.6.12
psycopg2-binary==2.9.5
python-dotenv==1.0.0